Move the heavy lifting out of the root module
